refactor and add comment
